LayoutTests/platform/win/accessibility/aria-switch-text-expected.txt

One
Two
Three
This tests that ARIA switches use accessible name computation and have no child text.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


widget.title is AXTitle: One
widget.description is AXDescription:
widget.childrenCount is 0
widget.title is AXTitle: Two
widget.description is AXDescription:
widget.childrenCount is 0
widget.title is AXTitle: Three
widget.description is AXDescription: foo
widget.childrenCount is 0
PASS successfullyParsed is true

TEST COMPLETE
